Added error handling to Giana.App.Cmd.

+++ src/app/cmd/Program.cs
@@ -5,7 +5,33 @@ using System.IO;
 using System.Linq;
 using static Giana.App.Shared.Actions;
 
+const string GitExePathEnvName = "GitExePath";
+
 var cmdLineArgs = Environment.GetCommandLineArgs().ToList();
+
+int idxHelp = Math.Max(cmdLineArgs.IndexOf("-h"), cmdLineArgs.IndexOf("--help"));
+if (idxHelp > 0)
+{
+  Console.WriteLine("usage: Giana.Cmd.App.[bat|sh] [(-q | --query-file) <filename>] [(-o | output-file) <filename>]");
+  Console.WriteLine();
+  Console.WriteLine("--query-file\tjson file with the query arguments. By default, query.json for the bin folder is used.");
+  Console.WriteLine("--output-file\tcsv filename for the results. By default, the results are written to the standard output.");
+
+  return;
+}
+
+string gitExePath = Environment.GetEnvironmentVariable(GitExePathEnvName);
+if (string.IsNullOrEmpty(gitExePath))
+{
+  Console.WriteLine($"environment variable '{GitExePathEnvName}' not found.");
+  return;
+}
+
+if (!File.Exists(gitExePath))
+{
+  Console.WriteLine($"File '{gitExePath}' defined with environment variable '{GitExePathEnvName}' not found.");
+  return;
+}
 
 string queryFilename = null;
 TextWriter outputWriter = null;
@@ -21,11 +47,25 @@ else
   queryFilename = Path.Combine(cwd, "query.json");
 }
 
+if (!File.Exists(queryFilename))
+{
+  Console.WriteLine($"File '{queryFilename}' defined with the command line argument '{cmdLineArgs[idxQueryFile]}' not found.");
+  return;
+}
+
 int idxOutputFile = Math.Max(cmdLineArgs.IndexOf("-o"), cmdLineArgs.IndexOf("--output-file"));
 if (idxOutputFile > 0 && cmdLineArgs.Count > idxOutputFile + 1)
 {
-  var outputStream = File.Open(cmdLineArgs[idxOutputFile + 1], FileMode.OpenOrCreate);
-  outputWriter = new StreamWriter(outputStream);
+  try
+  {
+    var outputStream = File.Open(cmdLineArgs[idxOutputFile + 1], FileMode.OpenOrCreate);
+    outputWriter = new StreamWriter(outputStream);
+  }
+  catch
+  {
+    Console.WriteLine($"Failed to create output file {cmdLineArgs[idxOutputFile + 1]} defined with the command line argument '{cmdLineArgs[idxOutputFile]}'.");
+    return;
+  }
 }
 else
 {
@@ -46,7 +86,7 @@ var routine = Calculations.CreateRoutine(query);
 
 routine.OutputWriter = outputWriter;
 
-var gitLogRecords = Execute(routine);
+var gitLogRecords = Execute(routine, () => gitExePath);
 
 outputWriter.Flush();
 outputWriter.Dispose();
